Extract opponent selection to separate class
* Create OpponentSelectionTest and move opponent selection tests to it
* Create OpponentSelector and move opponent selection methods to it
* Add Game.opponent_selector
