+++ spec/game_test.rb
@@ -35,20 +35,4 @@ class GameTest < Minitest::Test
   def test_implements_current_player
     assert_respond_to(@test_game, :current_player)
   end
-
-  def test_choose_human_opponent
-    $stdin = input_to_string_io('human')
-
-    @test_game.choose_opponent
-
-    assert_kind_of(Player, @test_game.opponent)
-  end
-
-  def test_choose_computer_opponent
-    $stdin = input_to_string_io('computer')
-
-    @test_game.choose_opponent
-
-    assert_kind_of(CompPlayer, @test_game.opponent)
-  end
 end
